+++ addin/src/functions/functions.js
@@ -1,25 +1,43 @@
-/* global CustomFunctions, document, Excel, Office */
+/* global Excel */
 
 /**
- * Get value for key
+ * Find the class that a cohort is in given a day and time.
  * @customfunction
- * @param key The key
- * @returns The value for the key.
+ * @param {string} cohort The cohort to search for
+ * @param {string} day The day of the week
+ * @param {string} timeslot The timeslot
+ * @returns {string} The class the cohort is in.
  */
-function getValueForKeyCF(key) {
-  return getValueForKey(key);
-}
+async function FindCohortClass(cohort, day, timeslot) {
+  // console.log("find cohort")
+  try {
+    return await Excel.run(async (context) => {
+      const sheet = context.workbook.worksheets.getItem("Schedule");
+      const range = sheet.getUsedRange();
+      range.load("values");
+      await context.sync();
 
-/**
- * Get value for key
- * @customfunction
- * @param key The key
- * @returns The value for the key.
- */
-function setValueForKeyCF(key, value) {
-  setValueForKey(key, value);
-  return "Stored key/value pair";
-}
+      const values = range.values;
+      const headerRow = values[0];
 
-CustomFunctions.associate("GETVALUEFORKEYCF", getValueForKeyCF);
-CustomFunctions.associate("SETVALUEFORKEYCF", setValueForKeyCF);
+      // Find column with matching day,timeslot
+      const columnIndex = headerRow.findIndex(
+        header => header === `${day}, ${timeslot}`
+      );
+
+      if (columnIndex === -1) return "Time not found";
+
+      // Find row with matching cohort name
+      for (let i = 1; i < values.length; i++) {
+        // console.log(`checking ${values[i][0]} for ${cohort}, from column ${columnIndex}`)
+        if (values[i][columnIndex] === cohort) {
+          return values[i][0];
+        }
+      }
+
+      return "No Class";
+    });
+  } catch (error) {
+    return "Error: " + error.message;
+  }
+}