+++ changelog.md
@@ -10,12 +10,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Add
 
 - data validation from the AllCohorts rule
-- the FindClassCohortIsIn function
-- make the panel smaller
 
 ## v0.2.0 - Unreleased
 
 ### Add
+- the FindCohortClass function, and docs
 
 ### Changed
 - made the panel contents smaller, added version